+++ snippets-source/components/import-effects.js
@@ -1,0 +1,2 @@
+import { Expand, Push, Reveal, Slide, Zoom } from '@progress/kendo-react-animation';
+import { Ripple } from '@progress/kendo-react-ripple';