+++ snippets-source/components/import.js
@@ -3,4 +3,7 @@ import { Button, ButtonGroup } from '@progress/kendo-react-buttons';
 import { DateInput, Calendar } from '@progress/kendo-react-dateinputs';
 import { DropDownList, ComboBox } from '@progress/kendo-react-dropdowns';
 import { PanelBar, PanelBarItem, TabStrip, TabStripTab } from '@progress/kendo-react-layout';
-import { process } from '@progress/kendo-data-query';
+import { NumericTextBox, Input, Switch } from '@progress/kendo-react-inputs';
+import { Dialog } from '@progress/kendo-react-dialogs';
+import { Popup } from '@progress/kendo-react-popup';
+import { Chart, Sparkline, StockChart } from '@progress/kendo-react-charts';