+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "kendo-react-snippets" extension will be documented i
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+- Updated import snippets for R2 release. Introduced new krc-import-common, krc-import-drawing, krc-import-effects, and krc-import-export statements. 
 
 ## [0.3.1] - 2018-03-16